🚀 Expose methods to get serializable public and private key (#27)
* ♻ Clean up spacing in Project files

Visual studio cleanup on project files

* 🚧 Create Serialization for Public Key and Trustee State

Create serializers that can be used for manipulating the public key and private key aka trustee state

* 🚀 Implement Trustee State and Public Key Export

Public Key and Trustee are now exported and used within tests so that any user can access them.

* 🚧 Disable Windows Test

This is temporary disabled due to waiting on PR's in C-Implementation. This should be removed later.