Create an Election  (#30)
* ♻ Refactor uint to int and string for baseHash

uint is just not friendly for new developers. Switch to int where possible. Use basehash as string where possible. Easier for developers to work with.

* 🚧 Update Serializers to use base64

Use base64 for byte conversion  as primary

* ✨ Create Election Constructor

Election constructor was made to quickly assemble all the details of an election with a single call for implementation into the api

* 🔧 Badge should be based on Package